Separated quads and triangle tests
